Add files via upload
Power BI is a great option for analyzing a table or any other data with images and charts easily. Here a table with some products and their sale report in different countries. We can analyze this data with many graphs and charts and identify the trends in societies and can find the total profit from the sale.
   The dashboard contains the charts and graphs about their total gross sale and profit of different products from different countries. This will help to identify the money that got for the company and the profit attain from the total sale.